Uniformize export key dialog
