+++ vector/src/main/java/im/vector/activity/signout/SignOutActivity.kt
@@ -19,14 +19,9 @@ package im.vector.activity.signout
 import android.arch.lifecycle.Observer
 import android.arch.lifecycle.ViewModelProviders
 import android.content.DialogInterface
-import android.support.design.widget.TextInputEditText
 import android.support.v4.content.ContextCompat
 import android.support.v7.app.AlertDialog
-import android.text.Editable
-import android.text.TextUtils
-import android.text.TextWatcher
 import android.view.View
-import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.core.view.isVisible
@@ -36,6 +31,7 @@ import im.vector.R
 import im.vector.activity.CommonActivityUtils
 import im.vector.activity.KeysBackupManageActivity
 import im.vector.activity.MXCActionBarActivity
+import im.vector.dialogs.ExportKeysDialog
 import im.vector.ui.themes.ThemeUtils
 import im.vector.util.PERMISSIONS_FOR_WRITING_FILES
 import im.vector.util.PERMISSION_REQUEST_CODE
@@ -43,6 +39,7 @@ import im.vector.util.allGranted
 import im.vector.util.checkPermissions
 import org.matrix.androidsdk.crypto.keysbackup.KeysBackupStateManager
 import org.matrix.androidsdk.rest.callback.SimpleApiCallback
+import org.matrix.androidsdk.rest.model.MatrixError
 
 class SignOutActivity : MXCActionBarActivity() {
     private lateinit var viewModel: SignOutViewModel
@@ -189,53 +186,38 @@ class SignOutActivity : MXCActionBarActivity() {
     }
 
     private fun doExportKeysToFile() {
-        val dialogLayout = layoutInflater.inflate(R.layout.dialog_export_e2e_keys, null)
-        val builder = AlertDialog.Builder(this)
-                .setTitle(R.string.encryption_export_room_keys)
-                .setView(dialogLayout)
+        ExportKeysDialog().show(this, object : ExportKeysDialog.ExportKeyDialogListener {
+            override fun onPassphrase(passphrase: String) {
+                showWaitingView()
 
-        val passPhrase1EditText = dialogLayout.findViewById<TextInputEditText>(R.id.dialog_e2e_keys_passphrase_edit_text)
-        val passPhrase2EditText = dialogLayout.findViewById<TextInputEditText>(R.id.dialog_e2e_keys_confirm_passphrase_edit_text)
-        val exportButton = dialogLayout.findViewById<Button>(R.id.dialog_e2e_keys_export_button)
-        val textWatcher = object : TextWatcher {
-            override fun beforeTextChanged(s: CharSequence, start: Int, count: Int, after: Int) {
+                CommonActivityUtils.exportKeys(mSession, passphrase, object : SimpleApiCallback<String>(this@SignOutActivity) {
+                    override fun onSuccess(filename: String) {
+                        hideWaitingView()
 
+                        AlertDialog.Builder(this@SignOutActivity)
+                                .setMessage(getString(R.string.encryption_export_saved_as, filename))
+                                .setCancelable(false)
+                                .setPositiveButton(R.string.ok, null)
+                                .show()
+                    }
+
+                    override fun onNetworkError(e: Exception) {
+                        super.onNetworkError(e)
+                        hideWaitingView()
+                    }
+
+                    override fun onMatrixError(e: MatrixError) {
+                        super.onMatrixError(e)
+                        hideWaitingView()
+                    }
+
+                    override fun onUnexpectedError(e: Exception) {
+                        super.onUnexpectedError(e)
+                        hideWaitingView()
+                    }
+                })
             }
-
-            override fun onTextChanged(s: CharSequence, start: Int, before: Int, count: Int) {
-                exportButton.isEnabled = !TextUtils.isEmpty(passPhrase1EditText.text) && TextUtils.equals(passPhrase1EditText.text, passPhrase2EditText.text)
-            }
-
-            override fun afterTextChanged(s: Editable) {
-
-            }
-        }
-
-        passPhrase1EditText.addTextChangedListener(textWatcher)
-        passPhrase2EditText.addTextChangedListener(textWatcher)
-
-        val exportDialog = builder.show()
-
-        exportButton.setOnClickListener {
-            showWaitingView()
-
-            CommonActivityUtils.exportKeys(mSession, passPhrase1EditText.text.toString(), object : SimpleApiCallback<String>(this) {
-
-                override fun onSuccess(filename: String) {
-                    hideWaitingView()
-
-                    viewModel.keysExportedToFile.value = true
-
-                    AlertDialog.Builder(this@SignOutActivity)
-                            .setMessage(getString(R.string.encryption_export_saved_as, filename))
-                            .setCancelable(false)
-                            .setPositiveButton(R.string.ok, null)
-                            .show()
-                }
-            })
-
-            exportDialog.dismiss()
-        }
+        })
     }
 
     override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {
